// src/app/auth/model/users.model.ts
export class Users {
  uid?: string;
  firstname?: string;
  lastname?: string;
  displayName?: string;
  email?: string;
  phone?: string;
  emailVerified?: boolean;
  accountType?: string;
  firstrun? : string;
  location? : string;
  photoUrl? : string;
  points? : number;
  messagecreated? : boolean;
  country? : string;
}